Fix the indentation file
Completely rewrite the indentation file by manually extending the
optional JSX indentation (of `mxw/vim-jsx`).

The general workflow of the current indentation algorithm is to check,
whether the current line is a `jsTemplateString` (which get indented by
one default level) and indenting available `cssDefinitions` by another
level. If no css definitions are available or the css definitions get
closed by the end of the line the current line is unindented by one
level.

There are still several edge-cased to be handled but for now this works
more satisfactory than the previous implementation.