Bypass overwritten `indentexpr` in jsx files
Using `vim-polyglot` sometimes only javascript but sometimes jsx related
plugins get loaded in no particular order (for my untrained eye). The
`vim-jsx` plugin in particular sets the `indentexpr` and `indentkeys`
directly (i.e. uses `=` instead of `+=`).

Linking the jsx filetype to the javascript file ensures, that this
plugin is loaded afterwards as well and the custom definitions and
indentation functions are loaded correctly.

Fingers crossed, that it does, at least.