(Hopefully) fix infinite recursion on jsx files
Both this and the `jsx.vim` plugin evaluate the previously stored
indent expression on certain cases. This leads to an infinite recursion
between the two, as both have each other as the previous indent
expression.

Wrapping `eval(b:js_ts_indent)` in a try-catch block should abort on
errors due to this (and fallback to c-indentation).