Fix nested js in jsFuncCall region
Previously in a styled `jsFuncCall` region (e.g. `styled(<here>)`), only
keyword items (e.g. "\k") was highlighted, this thus was excluding
certain js constructs like property access syntax and function calls.

This commit changes this to basically ignore the specifics inside of the
insides of the `jsFuncCall` and to transparently highlight this. It also
adds styles to display the issue as was proposed in #44.

This should fix #44.