Add support for 'othree/yajs'
[othree/yajs](https://github.com/othree/yajs.vim) uses a different base
classification for syntax elements than
[pangloss/vim-javascript](https://github.com/pangloss/vim-javascript)
on which this plugin was originally developed for.

The general approach was to extend the available syntax clusters with
the already defined styled-definition related items, or redefine
existing yajs syntax items to allow for our custom ones.

Also this commit adds the issue-related file to make sure the syntax
highlighting is improved.