Don't trigger previous omnifunc if it wasn't set
In some cases (if there was no previous omnifunc set), the b:prevofu
reference would contain styledComplete#CompleteSC causing an infinite
loop to occur. So, it might be a good idea to not save the previous
omnifunc if it was not set and then not try to call it later on.